requirements.txt
numpy~=2.2.6
scipy~=1.16.2
pandas~=2.3.3
matplotlib~=3.10.6
scikit-image~=0.25.2
nibabel~=5.3.2
SimpleITK~=2.5.2
opencv-python~=4.12.0.88
scipy
ipywidgets~=8.1.7
SimpleITK
matplotlib
opencv-python
ipywidgets
pytorch-ignite
torch~=2.8.0
monai~=1.5.1
tqdm~=4.67.1

#not as necessary, depends
jupyter
ipykernel

